+++ src/pronet/test_tcp_client/main.cpp
@@ -265,7 +265,7 @@ int main(int argc, char* argv[])
         if (server_ip != NULL && server_port > 0)
         {
             configInfo.tcpc_server_ip   = server_ip;
-            configInfo.tcpc_server_port = server_port;
+            configInfo.tcpc_server_port = (unsigned short)server_port;
         }
 
         if (local_ip != NULL)
